[SKP-189] refac : linking 파일 분리
